Refactoring of postgres module
